+++ packages/design-system/components/web/logos/ProjectLogo/index.tsx
@@ -9,7 +9,6 @@ import Image from 'next/image'
 import { useLogoOverrides } from '../../../../hooks/web'
 import { ProjectLogoProps } from './types'
 
-// TODO: move this to @baseapp-frontend/design-system/components/web/logos
 const ProjectLogo: FC<ProjectLogoProps> = ({ src, alt, width, height, priority, className }) => {
   const { logos } = useLogoOverrides()
   const { isSSR } = useSSR()